{test => driver/software}.WindowlessCanvas
This is the final step to make the `driver/software` package independent
from the `test` package.

As a side-effect, it also removes the `test` dependency from `tools/playground`.
