+++ test/canvas.go
@@ -7,17 +7,7 @@ import (
 	"fyne.io/fyne/v2/driver/software"
 )
 
-var dummyCanvas WindowlessCanvas
-
-// WindowlessCanvas provides functionality for a canvas to operate without a window
-type WindowlessCanvas interface {
-	fyne.Canvas
-
-	Padded() bool
-	Resize(fyne.Size)
-	SetPadded(bool)
-	SetScale(float32)
-}
+var dummyCanvas software.WindowlessCanvas
 
 // Canvas returns a reusable in-memory canvas used for testing
 func Canvas() fyne.Canvas {
@@ -30,13 +20,13 @@ func Canvas() fyne.Canvas {
 
 // NewCanvas returns a single use in-memory canvas used for testing.
 // This canvas has no painter so calls to Capture() will return a blank image.
-func NewCanvas() WindowlessCanvas {
+func NewCanvas() software.WindowlessCanvas {
 	return wrapCanvas(software.NewCanvasWithPainter(nil))
 }
 
 // NewCanvasWithPainter allows creation of an in-memory canvas with a specific painter.
 // The painter will be used to render in the Capture() call.
-func NewCanvasWithPainter(painter fynedriver.Painter) WindowlessCanvas {
+func NewCanvasWithPainter(painter fynedriver.Painter) software.WindowlessCanvas {
 	return wrapCanvas(software.NewCanvasWithPainter(painter))
 }
 
@@ -44,15 +34,15 @@ func NewCanvasWithPainter(painter fynedriver.Painter) WindowlessCanvas {
 // The painter will be used to render in the Capture() call.
 //
 // Since: 2.2
-func NewTransparentCanvasWithPainter(painter fynedriver.Painter) WindowlessCanvas {
+func NewTransparentCanvasWithPainter(painter fynedriver.Painter) software.WindowlessCanvas {
 	return wrapCanvas(software.NewTransparentCanvasWithPainter(painter))
 }
 
-func wrapCanvas(c WindowlessCanvas) *canvas {
+func wrapCanvas(c software.WindowlessCanvas) *canvas {
 	return &canvas{WindowlessCanvas: c}
 }
 
 type canvas struct {
-	WindowlessCanvas
+	software.WindowlessCanvas
 	hovered desktop.Hoverable
 }
