+++ tools/playground/softwarecanvas.go
@@ -2,10 +2,9 @@ package playground
 
 import (
 	"fyne.io/fyne/v2/driver/software"
-	"fyne.io/fyne/v2/test"
 )
 
 // NewSoftwareCanvas creates a new canvas in memory that can render without hardware support
-func NewSoftwareCanvas() test.WindowlessCanvas {
+func NewSoftwareCanvas() software.WindowlessCanvas {
 	return software.NewCanvas()
 }
