+++ widget/radio_group_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/driver/software"
 	"fyne.io/fyne/v2/layout"
 	"fyne.io/fyne/v2/test"
 	"fyne.io/fyne/v2/widget"
@@ -18,7 +19,7 @@ func TestRadioGroup_FocusRendering(t *testing.T) {
 		defer window.Close()
 		window.Resize(radio.MinSize().Max(fyne.NewSize(150, 200)))
 
-		canvas := window.Canvas().(test.WindowlessCanvas)
+		canvas := window.Canvas().(software.WindowlessCanvas)
 		test.AssertRendersToMarkup(t, "radio_group/focus_none_focused_none_selected.xml", canvas)
 		canvas.FocusNext()
 		test.AssertRendersToMarkup(t, "radio_group/focus_a_focused_none_selected.xml", canvas)
@@ -43,7 +44,7 @@ func TestRadioGroup_FocusRendering(t *testing.T) {
 		defer window.Close()
 		window.Resize(radio.MinSize().Max(fyne.NewSize(150, 200)))
 
-		canvas := window.Canvas().(test.WindowlessCanvas)
+		canvas := window.Canvas().(software.WindowlessCanvas)
 		canvas.FocusNext()
 		radio.Disable()
 		test.AssertRendersToMarkup(t, "radio_group/focus_disabled_none_selected.xml", canvas)
@@ -57,7 +58,7 @@ func TestRadioGroup_FocusRendering(t *testing.T) {
 		defer window.Close()
 		window.Resize(radio.MinSize().Max(fyne.NewSize(150, 200)))
 
-		canvas := window.Canvas().(test.WindowlessCanvas)
+		canvas := window.Canvas().(software.WindowlessCanvas)
 		radio.Disable()
 		test.AssertRendersToMarkup(t, "radio_group/disabled_none_selected.xml", canvas)
 		radio.Append("Option C")
@@ -170,7 +171,7 @@ func TestRadioGroup_ToggleSelectionWithSpaceKey(t *testing.T) {
 
 	assert.Equal(t, "", radio.Selected)
 
-	canvas := window.Canvas().(test.WindowlessCanvas)
+	canvas := window.Canvas().(software.WindowlessCanvas)
 	canvas.FocusNext()
 	canvas.FocusNext()
 	test.Type(canvas.Focused(), " ")
